Refactor: Context 범위 수정
